Removed references to outdated TinkerPop versions
Removed references to TinkerPop versions that are now basically referring to features far enough in the past that they simply reflect the current way of doing things. We will probably keep this approach to mentioning versions but base it from a more current version of TinkerPop. Did not remove references to versions in the "Introduction" where it was introducing specific versions and their new features as that carries a history that makes sense in that context.